fix: BaseEntity 펀딩모듈로 이동
