--- bc/funding/src/main/java/domain/WishlistItem.java
@@ -1,23 +1,0 @@
-package domain;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.NoArgsConstructor;
-
-@Entity
-@NoArgsConstructor
-public class WishlistItem {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private Wishlist wishlist;
-
-    private Product product;
-
-    private WishlistStatus status;
-
-    // TODO: baseEntity 상속
-}
